requirements.txt
Django==3.1.4
django-guardian==2.4.0
djangorestframework-guardian==0.3.0
djangorestframework==3.12.2
psycopg2-binary==2.9.7
redis==3.5.0
PyJWT==2.8.0
celery==5.3.1
black==23.12.0
flake8==7.1.1
isort==5.13.2